requirements.txt
duckdb==1.5.2
pandas==2.2.3
pyarrow==18.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
